Renaming cxx to cpp. Changes according to renaming of DetectorDescription.
