+++ Detector/DetFCChhBaseline/compact/FCChh_BeamTube.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<lccdd xmlns:compact="http://www.lcsim.org/schemas/compact/1.0" 
-       xmlns:xs="http://www.w3.org/2001/XMLSchema" 
+<lccdd xmlns:compact="http://www.lcsim.org/schemas/compact/1.0"
+       xmlns:xs="http://www.w3.org/2001/XMLSchema"
        xs:noNamespaceSchemaLocation="http://www.lcsim.org/schemas/compact/1.0/compact.xsd">
-  
+
   <info name="FCC"
         title="FCC Beam pipe"
         author="Clement Helsens"
@@ -12,27 +12,9 @@
     <comment>A beam pipe for FCC detector</comment>
   </info>
 
-
-  <define>
-        <constant name="CentralBeamTubeZ" value="800.0*cm"/>
-        <constant name="CentralBeamTubeRmax" value="2.1*cm"/>
-        <constant name="CentralBeamTubeRmin" value="2.*cm"/>
-
-        <constant name="ForwardBeamTubeRmin1" value="2.*cm"/>
-        <constant name="ForwardBeamTubeRmax1" value="2.1*cm"/>
-        <constant name="ForwardBeamTubeRmin2" value="7.8*cm"/>
-        <constant name="ForwardBeamTubeRmax2" value="7.9*cm"/>
-        <constant name="ForwardBeamTubeZ" value="15600*cm"/>
-        <constant name="ForwardBeamTubeZoffset" value="800*cm"/>
-
-    </define>
-
-
   <display>
     <vis name="BeamPipeVis" r="0.2" g="0.7" b="0.3" showDaughters="false" visible="true"/>
   </display>
-  
-  
 
   <detectors>
     <detector name="CentralBeamTube" type="SimpleCylinder"  vis="BeamPipeVis">
@@ -44,19 +26,15 @@
     <detector name="ForwardBeamTube1" type="SimpleCone"  vis="BeamPipeVis">
       <status id="1"/>
       <comment>Forward Be Beampipe positive z</comment>
-      <dimensions rmin1="ForwardBeamTubeRmin1" rmax1="ForwardBeamTubeRmax1" rmin2="ForwardBeamTubeRmin2" rmax2="ForwardBeamTubeRmax2" dz="ForwardBeamTubeZ" z_offset="ForwardBeamTubeZoffset+ForwardBeamTubeZ"  material="Beryllium"/>  
+      <dimensions rmin1="ForwardBeamTubeRmin1" rmax1="ForwardBeamTubeRmax1" rmin2="ForwardBeamTubeRmin2" rmax2="ForwardBeamTubeRmax2" dz="ForwardBeamTubeZ" z_offset="ForwardBeamTubeZoffset+ForwardBeamTubeZ"  material="Beryllium"/>
     </detector>
 
     <detector name="ForwardBeamTube2" type="SimpleCone"  vis="BeamPipeVis">
       <status id="2"/>
       <comment>Forward Be Beampipe negative z</comment>
-      <dimensions rmin1="ForwardBeamTubeRmin1" rmax1="ForwardBeamTubeRmax1" rmin2="ForwardBeamTubeRmin2" rmax2="ForwardBeamTubeRmax2" dz="-ForwardBeamTubeZ" z_offset="-ForwardBeamTubeZoffset-ForwardBeamTubeZ" material="Beryllium"/>  
+      <dimensions rmin1="ForwardBeamTubeRmin1" rmax1="ForwardBeamTubeRmax1" rmin2="ForwardBeamTubeRmin2" rmax2="ForwardBeamTubeRmax2" dz="-ForwardBeamTubeZ" z_offset="-ForwardBeamTubeZoffset-ForwardBeamTubeZ" material="Beryllium"/>
     </detector>
-
-
-
-
   </detectors>
-  
+
 </lccdd>
