+++ Sim/SimG4Components/CMakeLists.txt
@@ -4,7 +4,7 @@
 gaudi_subdir(SimG4Components v1r0)
 
 # this declaration will not be needed in the future
-gaudi_depends_on_subdirs(GaudiAlg FWCore Sim/SimG4Common DetectorDescription/DetDesInterfaces Sim/SimG4Interface)
+gaudi_depends_on_subdirs(GaudiAlg FWCore Sim/SimG4Common Detector/DetInterface Sim/SimG4Interface)
 
 find_package(Geant4)
 include(${Geant4_USE_FILE})
